Introduce couple of storage-only types. Cleanup and fix warnings. No functionality change.

+++ assembler/src/debruijn/graphio.hpp
@@ -253,7 +253,7 @@ class DataPrinter {
                         fprintf(file, "%zu %zu %.2f %.2f %.2f .\n",
                                 e1.int_id(),
                                 e2.int_id(),
-                                point.d, point.weight, point.var);
+                                (double)point.d, (double)point.weight, (double)point.var);
                     }
             }
         }
